[HBASE-6423] [0.89-fb] change hbase block multiplier to support float.
Author: aaiyer

Summary:
Currently this is set to long. So, we have to set it to 2.0 at least.
Changing it to float will allow us better control.

Test Plan: mr unit test.

Reviewers: kannan, kranganathan

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D546696

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1378339 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/hadoop/hbase/regionserver/HRegion.java
@@ -527,8 +527,8 @@ public class HRegion implements HeapSize {
     }
     this.disableWAL = regionInfo.getTableDesc().isWALDisabled();
     this.memstoreFlushSize = flushSize;
-    this.blockingMemStoreSize = this.memstoreFlushSize *
-      conf.getLong(HConstants.HREGION_MEMSTORE_BLOCK_MULTIPLIER, 2);
+    this.blockingMemStoreSize = (long)(this.memstoreFlushSize *
+      conf.getFloat(HConstants.HREGION_MEMSTORE_BLOCK_MULTIPLIER, 2));
     this.waitOnMemstoreBlock =
         conf.getBoolean(HConstants.HREGION_MEMSTORE_WAIT_ON_BLOCK, true);
     this.scannerReadPoints = new ConcurrentHashMap<RegionScanner, Long>();
